Holy Cow is PowerShell busted (#349)
* We broke powershell when we removed all the dll's from chef. We also needed to update all the things.

+++ .expeditor/build_gems.ps1
@@ -11,24 +11,47 @@
 
 $ErrorActionPreference = "Stop"
 
+Write-Output "Are my Hab environment variables set correctly?"
+if ([string]::IsNullOrEmpty($env:HAB_AUTH_TOKEN)){
+  Write-Output ( "=" * 80 )
+  Write-Error "HAB_AUTH_TOKEN is not set! Please set it before running this script."
+  Write-Output ( "=" * 80 )
+  exit 1
+}
+else {
+  Write-Output "HAB_AUTH_TOKEN is set correctly."
+}
+
 Write-Output "--- :ruby: Removing existing Ruby instances"
 
 $rubies = Get-ChildItem -Path "C:\ruby*"
 foreach ($ruby in $rubies){
   Remove-Item -LiteralPath $ruby.FullName -Recurse -Force -ErrorAction SilentlyContinue
 }
+
 Write-Output "`r"
 
 # Need to set this variable to keep the build from failing while trying to resolve nonsense sdk paths
 $env:MSBuildEnableWorkloadResolver = "false"
 
-# setting the channel in this way gets access to the LTS channel and falls back to stable if the plan doesn't live there.
-Write-Output "--- :shovel: Setting the BLDR Channel to LTS"
-Write-Output "`r"
-
-Write-Output "--- :screwdriver: Installing Habitat via Choco"
-choco install habitat -y
-if (-not $?) { throw "unable to install Habitat"}
+Write-Output "--- :screwdriver: Installing Habitat via Github"
+try {
+    [Version]$hab_version = (hab --version).split(" ")[1].split("/")[0]
+    if ($hab_version -lt [Version]"0.85.0" ) {
+        Write-Host "--- :habicat: Installing the version of Habitat required"
+        Set-ExecutionPolicy Bypass -Scope Process -Force
+        Invoke-Expression ((New-Object System.Net.WebClient).DownloadString('https://raw.githubusercontent.com/habitat-sh/habitat/main/components/hab/install.ps1'))
+        if (-not $?) { throw "Hab version is older than 0.85 and could not update it." }
+    } else {
+        Write-Host "--- :habicat: :thumbsup: Minimum required version of Habitat already installed"
+    }
+}
+catch {
+    # This install fails if Hab isn't on the path when we check for the version. This ensures it is installed
+    Write-Host "--- :habicat: Installing the version of Habitat required"
+    Set-ExecutionPolicy Bypass -Scope Process -Force
+    Invoke-Expression ((New-Object System.Net.WebClient).DownloadString('https://raw.githubusercontent.com/habitat-sh/habitat/main/components/hab/install.ps1'))
+}
 Write-Output "`r"
 
 Write-Output "--- :screwdriver: Installing the latest Chef-Client"
@@ -40,6 +63,11 @@ Write-Output "--- :chopsticks: Refreshing the build environment to pick up Hab b
 Import-Module $env:ChocolateyInstall\helpers\chocolateyProfile.psm1
 refreshenv
 $env:Path = [System.Environment]::GetEnvironmentVariable("Path", "Machine") + ";" + [System.Environment]::GetEnvironmentVariable("Path", "User") + ";c:\opscode\chef\embedded\bin"
+# We are being really aggressive about setting the bldr channel because some process keeps changing it to 'stable' which breaks our build
+$env:HAB_BLDR_CHANNEL = "base-2025"
+$env:HAB_ORIGIN = "chef"
+$env:HAB_LICENSE = "accept-no-persist"
+[System.Environment]::SetEnvironmentVariable("HAB_BLDR_CHANNEL", "base-2025", "Process")
 Write-Output "`r"
 
 Write-Output "--- :building_construction: Correcting a gem build problem, moving header files around"
@@ -53,7 +81,7 @@ Copy-Item $parent_folder -Destination $child_folder -ErrorAction Continue
 Write-Output "`r"
 
 Write-Output "--- :construction: Setting up Habitat to build PowerShell DLL's"
-$env:HAB_ORIGIN = "core"
+$env:HAB_ORIGIN = "chef"
 $env:HAB_LICENSE= "accept-no-persist"
 $env:FORCE_FFI_YAJL="ext"
 if (Test-Path -PathType leaf "/hab/cache/keys/core-*.sig.key") {
@@ -72,7 +100,8 @@ Write-Output "`r"
 
 
 Write-Output "--- :construction: Building 64-bit PowerShell DLLs"
-hab pkg build Habitat
+$env:HAB_BLDR_CHANNEL = "base-2025"
+hab pkg build habitat
 if (-not $?) { throw "unable to build"}
 Write-Output "`r"
 
@@ -88,7 +117,7 @@ if (-not $?) { throw "unable to install this build"}
 Write-Output "`r"
 
 Write-Output "--- :hammer_and_wrench: Capturing the x64 installation path"
-$x64 = hab pkg path core/chef-powershell-shim
+$x64 = hab pkg path chef/chef-powershell-shim
 Write-Output "Hab thinks it installed my 64-bit dlls here : $x64"
 Test-Path -Path $x64
 Write-Output "`r"
@@ -169,7 +198,12 @@ if (Test-Path $($parent_folder + "\chef.powershell.dll")){
 }
 Write-Output "`r"
 
-Write-Output "--- :point_right: finally verifying the gem code (chefstyle, spellcheck, spec)"
+Write-Output "--- :building_construction: It looks like I need to install chef-client"
+# as of this build, chef -client was not in base-2025 or stable so we grabbed this version directly
+hab pkg install chef/chef-infra-client/18.8.50/20251022232751 --channel "chef-chef-chef-18-habitat-build"
+Write-Output "`r"
+
+Write-Output "--- :point_right: finally verifying the gem code (cookstyle, spellcheck, spec)"
 bundle update
 bundle exec rake gem_check
 if (-not $?) { throw "Bundle Gem failed"}
